refactor(commands): isolate option semantics
